--- builtinstest.py
@@ -1,1 +1,0 @@
-print type(__builtins__)
