--- filetest.py
@@ -1,3 +1,0 @@
-print __file__
-print dir()
-print __package__
