--- inlinetest.py
@@ -1,4 +1,0 @@
-def test(*args):
-    print 'args: ', args
-    
-$test
