--- test.py
@@ -1,4 +1,0 @@
-def test():
-    x = 10
-def test2():
-    y = 10
