--- ttest.py
@@ -1,8 +1,0 @@
-def test(test_arg,
-         test_fail={"wow_lame" : 'yeah srsly'},
-         test_fail2= "testlol:\n    :\n\t",
-         *args,
-         **kwargs):
-    print 'this is a test function'
-    return None
-
